feat: add utility functions for API calls and list management
